[Refactor] 이미지개선 함수, 및 작성시 이미지 개선
